UPDATE restore Ubuntu configuration

+++ restore_linux_config.sh
@@ -1,15 +1,18 @@
 #!/bin/bash
 
-os=$(uname -o)
-if [ "$os == 'GNU/Linux'" ] && [ "$(uname -r | grep MANJARO)" ]; then
+COMMON_PACKAGES=(git terminator vlc aria2 tree bleachbit vim transmission-cli transmission-gtk htop gparted preload zsh)
+BROWSER_PACKAGES=(chromium-browser)
+
+CURRENT_OS=$(uname -o)
+if [ "$CURRENT_OS == 'GNU/Linux'" ] && [ "$(uname -r | grep MANJARO)" ]; then
     read -rp 'Do you want install your softwares ? Y or N ' answer
     if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
 	sudo pacman -Rsc hexchat empathy brasero evolution ms-office-online \
 		jre8-openjdk steam-manjaro lollypop || \
 	    sudo pacman -Syu && sudo pacman-mirrors -c France && \
-	    sudo pacman -S base-devel git terminator vlc aria2 tree tk \
-	    bleachbit vim transmission-cli transmission-gtk lldb valgrind \
-	    zsh clang ccache clamav python-pip yaourt fuse-exfat util-linux \
+	    sudo pacman -S $COMMON_PACKAGES[@] && \
+	    sudo pacman -S base-devel tk lldb valgrind \
+	    clang ccache clamav python-pip yaourt fuse-exfat util-linux \
 	    exfat-utils preload xorg-xprop p7zip unrar tar rsync arj \
 	    cabextract rpmextract exfat-utils fuse-exfat a52dec faac faad2 \
 	    flac jasper lame libdca libdv gst-libav libmad libmpeg2 libtheora \
@@ -24,7 +27,7 @@ if [ "$os == 'GNU/Linux'" ] && [ "$(uname -r | grep MANJARO)" ]; then
 
     read -rp 'Do you want install your AUR softwares ? Y or N ' answer
     if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
-	yaourt -S spotify slack-desktop \
+	yaourt -S spotify slack-desktop vscodium-bin \
 	docker docker-compose docker-machine \
 	virtualbox powerpill
     fi
@@ -39,13 +42,48 @@ if [ "$os == 'GNU/Linux'" ] && [ "$(uname -r | grep MANJARO)" ]; then
 elif [ "$(uname -a | grep Ubuntu)" ]; then
     read -rp 'Do you want install your softwares ? Y or N ' answer
     if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
-	sudo apt update && sudo apt upgrade && \
-	sudo apt install vlc gparted bleachbit transmission-gtk vim \
-	    preload gcc build-essential cmake python3-dev \
+	sudo apt-add-repository ppa:apt-fast/stable -y && sudo apt-get update && sudo apt-get -y install apt-fast
+	sudo apt-fast update && sudo apt-fast -y upgrade && \
+	sudo apt-fast -y install $COMMON_PACKAGES[@] \
+	    gcc build-essential cmake python3-dev \
 	    ubuntu-restricted-extras simple-scan gnome-tweak-tool exfat-fuse \
-	    exfat-utils && \
+	    exfat-utils
+    fi
+
+    read -rp 'Do you want install your Firefox Developer Edition ? Y or N ' answer
+    if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
+	cd ~/Downloads/ && \
+	aria2c https://download.mozilla.org/\?product\=firefox-devedition-latest-ssl\&os\=linux64\&lang\=en-US && \
+	tar xjf firefox-*.tar.bz2 && \
+	sudo mv firefox /opt/firefox-de && \
+	sudo chown -R $USER /opt/firefox-de
+	echo "Need to configure PATH 'export PATH=/opt/firefox-d/firefox:PATH' and MOVE .desktop in ~/.local/share/application"
+    fi
+
+    read -rp 'Do you want install snap packages ? Y or N ' answer
+    if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
 	sudo snap install slack --classic && \
-	sudo snap spotify telegram htop
+	sudo snap istall spotify codium
+    fi
+
+    read -rp 'Do you want install Docker Engine ? Y or N ' answer
+    if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
+	sudo apt-fast -y install \
+             apt-transport-https \
+             ca-certificates \
+             curl \
+             gnupg-agent \
+             software-properties-common && \
+	curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add - && \
+	sudo apt-key fingerprint 0EBFCD88 && \
+	sudo add-apt-repository \
+             "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+             $(lsb_release -cs) \
+             stable" && \
+	sudo apt-fast update && \
+	sudo apt-fast -y install docker-ce docker-ce-cli containerd.io && \
+	sudo usermod -aG docker $LOGNAME && \
+	docker run hello-world
     fi
 fi
 
@@ -65,7 +103,7 @@ if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
     cd "$HOME" || return
 fi
 
-if [ "$os == 'GNU/Linux'" ] ; then
+if [ "$CURRENT_OS == 'GNU/Linux'" ] ; then
     read -rp 'Do you want install Anaconda ? Y or N ' answer
     if [ "$answer" = 'Y' ] || [ "$answer" = 'y' ]; then
 	wget --output-document=/tmp/anaconda.sh https://repo.continuum.io/archive/Anaconda3-2018.12-Linux-x86_64.sh /tmp/anaconda.sh
